Keep one list of the positions a type is written in
The thirteen positions the specification names were written out inside the one
sweep that had needed them. A second sweep over the same positions has to write
them out again, and the copy is a copy of the specification's table that nothing
holds against the other copies: a position added to the language is then covered
by whichever sweeps someone remembered to update, and a position that quietly
stopped being one is still swept by the sweeps that were not.

So the list is `TypePositions`, and a sweep reads it. What each sweep asks of a
cell stays its own — the delimiter sweep still crosses the positions with the type
forms and pins how many cells refuse — and which cells there are is one answer.
